requirements.txt
canvasvg==1.0.5
fonttools==4.37.3
keyboard==0.13.5
Pillow==9.2.0
pynput==1.7.6
matplotlib==3.6.0